Розділив код, що вже був на hpp та срр файли
